main.py
import os
import json
import asyncio
import configparser
from pathlib import Path
from telegram import Update
from telegram.ext import Updater, CommandHandler, CallbackContext
from quotexapi.stable_api import Quotex
import pandas as pd
import ta

# Load configuration
def load_config():
    config_path = Path("settings/config.ini")
    if not config_path.exists():
        config_path.parent.mkdir(exist_ok=True, parents=True)
        return {}
    
    config = configparser.ConfigParser()
    config.read(config_path, encoding="utf-8")
    return dict(config.items("settings"))

def save_config(email, password, email_pass, user_data_dir):
    config_path = Path("settings/config.ini")
    config = configparser.ConfigParser()
    config["settings"] = {
        "email": email,
        "password": password,
        "email_pass": email_pass,
        "user_data_dir": user_data_dir
    }
    with open(config_path, 'w') as configfile:
        config.write(configfile)

# Initialize Quotex client
def init_quotex_client(config):
    email = config.get("email")
    password = config.get("password")
    email_pass = config.get("email_pass")
    user_data_dir = config.get("user_data_dir")
    
    client = Quotex(
        email=email,
        password=password,
        email_pass=email_pass,
        user_data_dir=Path(user_data_dir)
    )
    return client

# Telegram bot commands
async def start(update: Update, context: CallbackContext):
    user = update.effective_user
    config = load_config()
    
    if not config:
        await update.message.reply_text(
            "Welcome to the Quotex bot!\n"
            "Please provide your Quotex account details:\n"
            "Use the format: /set_config <email> <password> <email_pass> <user_data_dir>"
        )
    else:
        # Reset the bot configuration
        config.clear()
        await update.message.reply_text(
            "Configuration has been reset. Please provide your Quotex account details:\n"
            "Use the format: /set_config <email> <password> <email_pass> <user_data_dir>"
        )

async def set_config(update: Update, context: CallbackContext):
    args = context.args
    if len(args) != 4:
        await update.message.reply_text(
            "Usage: /set_config <email> <password> <email_pass> <user_data_dir>"
        )
        return
    
    email, password, email_pass, user_data_dir = args
    save_config(email, password, email_pass, user_data_dir)
    await update.message.reply_text("Configuration updated successfully!")

async def help_command(update: Update, context: CallbackContext):
    await update.message.reply_text(
        "Available commands:\n"
        "/start - Start the bot and reset configuration\n"
        "/set_config <email> <password> <email_pass> <user_data_dir> - Set or update account configuration\n"
        "/help - Get help\n"
        "/trade - Start a trading session\n"
        "/status - Check account status\n"
        "/toggle_auto_trade - Toggle auto trading mode\n"
        "/set_demo_trade <amount> - Set demo trading mode and amount"
    )

async def toggle_auto_trade(update: Update, context: CallbackContext):
    user_data = context.user_data
    user_data['auto_trade'] = not user_data.get('auto_trade', False)
    status = 'enabled' if user_data['auto_trade'] else 'disabled'
    await update.message.reply_text(f"Auto trading has been {status}.")

async def set_demo_trade(update: Update, context: CallbackContext):
    user_data = context.user_data
    try:
        amount = float(context.args[0])
        user_data['demo_trade'] = True
        user_data['demo_amount'] = amount
        await update.message.reply_text(f"Demo trading set with amount {amount}.")
    except (IndexError, ValueError):
        await update.message.reply_text("Usage: /set_demo_trade <amount>")

async def trade(update: Update, context: CallbackContext):
    user_data = context.user_data
    auto_trade = user_data.get('auto_trade', False)
    demo_trade = user_data.get('demo_trade', False)
    demo_amount = user_data.get('demo_amount', 0)

    if demo_trade:
        await update.message.reply_text(f"Demo trading with amount {demo_amount}.")
    else:
        await update.message.reply_text("Live trading started.")

    if auto_trade:
        await update.message.reply_text("Auto trading is enabled.")
        # Execute trades automatically
    else:
        await update.message.reply_text("Auto trading is disabled. Providing signals only.")
        # Provide trading signals

async def status(update: Update, context: CallbackContext):
    await update.message.reply_text("Account status: ... (implement status checking)")

def calculate_indicators(data):
    data['MA5'] = data['close'].rolling(window=5).mean()
    data['MA20'] = data['close'].rolling(window=20).mean()
    data['BB_upper'] = ta.volatility.BollingerBands(close=data['close'], window=20, window_dev=2).bollinger_hband()
    data['BB_lower'] = ta.volatility.BollingerBands(close=data['close'], window=20, window_dev=2).bollinger_lband()
    data['MACD'] = ta.trend.MACD(close=data['close'], window_slow=26, window_fast=12, window_sign=9).macd()
    data['RSI'] = ta.momentum.RSIIndicator(close=data['close'], window=14).rsi()
    return data

def main():
    config = load_config()
    client = init_quotex_client(config)

    updater = Updater("YOUR_TELEGRAM_BOT_TOKEN")
    dp = updater.dispatcher

    dp.add_handler(CommandHandler("start", start))
    dp.add_handler(CommandHandler("set_config", set_config))
    dp.add_handler(CommandHandler("help", help_command))
    dp.add_handler(CommandHandler("toggle_auto_trade", toggle_auto_trade))
    dp.add_handler(CommandHandler("set_demo_trade", set_demo_trade))
    dp.add_handler(CommandHandler("trade", trade))
    dp.add_handler(CommandHandler("status", status))

    updater.start_polling()
    updater.idle()

if __name__ == '__main__':
    main()