requirements.txt
python-telegram-bot
quotexapi
pandas
ta